requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.0
requests==2.28.2
seaborn==0.12.2
scikit_learn==1.2.2
streamlit==1.20.0
lightgbm==3.3.5
shap==0.41.0
